// bubbles-ui/src/gql/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
  ID: { input: number; output: number; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /**
   * Define a Relay Cursor type:
   * https://relay.dev/graphql/connections.htm#sec-Cursor
   */
  Cursor: { input: any; output: any; }
  Time: { input: string; output: string; }
};
// ...
export type CreateOrderCustomItemInput = {
  masterCustomItemID: Scalars['ID']['input'];
  quantity: Scalars['Int']['input'];
  selectedCustomItemIDs?: InputMaybe<Array<Scalars['ID']['input']>>;
  selectedCustomItems: Array<CreateSelectedCustomItemInput>;
};

/**
 * CreateOrderInput is used for create Order object.
 * Input was generated by ent.
 */
export type CreateOrderInput = {
  customItemIDs?: InputMaybe<Array<Scalars['ID']['input']>>;
  customItems: Array<CreateOrderCustomItemInput>;
  itemIDs?: InputMaybe<Array<Scalars['ID']['input']>>;
  items: Array<CreateOrderItemInput>;
  state?: InputMaybe<OrderState>;
};

/**
 * CreateOrderItemInput is used for create OrderItem object.
 * Input was generated by ent.
 */
export type CreateOrderItemInput = {
  itemID: Scalars['ID']['input'];
  quantity: Scalars['Int']['input'];
};

/**
 * CreateSelectedCustomItemInput is used for create SelectedCustomItem object.
 * Input was generated by ent.
 */
export type CreateSelectedCustomItemInput = {
  customItemID: Scalars['ID']['input'];
  selectedVariantIDs: Array<Scalars['ID']['input']>;
};

export type CustomItem = Node & {
  __typename?: 'CustomItem';
  /** if false then one or more variants can be selected at once */
  allowOnlyOne: Scalars['Boolean']['output'];
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  /** The id of the previous to selectable custom item */
  prev?: Maybe<Scalars['Int']['output']>;
  variants: Array<Item>;
};
// ...
export type Item = Node & {
  __typename?: 'Item';
  id: Scalars['ID']['output'];
  /** an url; eg.: https://example.com/cheesecake.png */
  image: Scalars['String']['output'];
  /** weather the item is (physically) in stock */
  inStock: Scalars['Boolean']['output'];
  /** eg.: cheesecake */
  name: Scalars['String']['output'];
  /** Custom text, which will be displayed */
  notes: Scalars['String']['output'];
  price: Scalars['Float']['output'];
};
// ...
export type Node = {
  /** The id of the object. */
  id: Scalars['ID']['output'];
};
// ...
export enum OrderState {
  Cancelled = 'cancelled',
  Compleated = 'compleated',
  Created = 'created',
  Pending = 'pending'
}